backend/requirements.txt
# --- ML & Image Processing ---
numpy<2.0.0
opencv-python-headless
Pillow
python-chess
tensorflow>=2.18.0

# --- FastAPI & Server ---
fastapi
uvicorn[standard]
python-multipart
pydantic-settings

# --- Database & ORM ---
sqlalchemy
asyncpg
alembic

# --- Authentication ---
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2

email-validator
python-multipart
boto3